Refactors and renaming
